Unit test Controllers

+++ src/Controller/TaskController.php
@@ -65,6 +65,7 @@ use Symfony\Component\Routing\Annotation\Route;
     public function editAction(Task $task, Request $request, EntityManagerInterface $entityManager): Response
     {
         if (!$this->userTaskPermissions::isAdmin($this->getUser()) && !$this->userTaskPermissions::isOwner($this->getUser(), $task)) {
+            $this->addFlash('error', 'Vous ne pouvez pas modifier cette tâche.');
             return $this->redirectToRoute('task_list');
         }
 
@@ -91,6 +92,7 @@ use Symfony\Component\Routing\Annotation\Route;
     public function toggleTaskAction(Task $task, EntityManagerInterface $entityManager): Response
     {
         if (!$this->userTaskPermissions->isAdmin($this->getUser()) && !$this->userTaskPermissions->isOwner($this->getUser(), $task)) {
+            $this->addFlash('error', 'Vous ne pouvez pas modifier cette tâche.');
             return $this->redirectToRoute('task_list');
         }
 
@@ -106,7 +108,10 @@ use Symfony\Component\Routing\Annotation\Route;
     #[Route('/tasks/{id}/delete', name: 'task_delete', methods: ['GET', 'POST'])]
     public function deleteTaskAction(Task $task, EntityManagerInterface $entityManager): Response
     {
-
+        if (!$this->userTaskPermissions::isAdmin($this->getUser()) && !$this->userTaskPermissions::isOwner($this->getUser(), $task)) {
+            $this->addFlash('error', 'Vous ne pouvez pas supprimer cette tâche.');
+            return $this->redirectToRoute('task_list');
+        }
         $entityManager->remove($task);
         $entityManager->flush();
